trying to run migrations for static analysis - should share with syntax
