+++ server/src/postgres/queries/queryFileSyntaxAnalysis.ts
@@ -1,6 +1,3 @@
-import fs from "fs/promises"
-import glob from "glob-promise"
-import path from "path"
 import { DatabaseError } from "pg"
 import {
   Diagnostic, DiagnosticSeverity, Logger, uinteger,
@@ -8,22 +5,18 @@ import {
 import { TextDocument } from "vscode-languageserver-textdocument"
 
 import { MigrationError } from "@/errors"
-import { PostgresClient, PostgresPool } from "@/postgres"
+import { PostgresPool } from "@/postgres"
 import {
   getQueryParameterInfo, QueryParameterInfo,
   sanitizeFileWithQueryParameters,
 } from "@/postgres/parameters"
-import { MigrationsSettings, Settings, StatementsSettings } from "@/settings"
-import { asyncFlatMap } from "@/utilities/functool"
+import { runMigration } from "@/postgres/queries/migrations"
+import { Settings, StatementsSettings } from "@/settings"
 import { neverReach } from "@/utilities/neverReach"
+import {
+  BEGIN_RE, COMMIT_RE, DISABLE_STATEMENT_VALIDATION_RE, ROLLBACK_RE, SQL_COMMENT_RE,
+} from "@/utilities/regex"
 import { getCurrentLineFromIndex, getTextAllRange } from "@/utilities/text"
-
-const SQL_COMMENT_RE = /\/\*[\s\S]*?\*\/|([^:]|^)--.*$/gm
-const BEGIN_RE = /^([\s]*begin[\s]*;)/igm
-const COMMIT_RE = /^([\s]*commit[\s]*;)/igm
-const ROLLBACK_RE = /^([\s]*rollback[\s]*;)/igm
-
-const DISABLE_STATEMENT_VALIDATION_RE = /^ *-- +plpgsql-language-server:disable *$/m
 
 export type SyntaxAnalysisOptions = {
   isComplete: boolean
@@ -139,103 +132,6 @@ function statementError(
     range,
     message,
   }
-}
-
-async function runMigration(
-  pgClient: PostgresClient,
-  document: TextDocument,
-  migrations: MigrationsSettings,
-  logger: Logger,
-): Promise<void> {
-  const upMigrationFiles = (
-    await asyncFlatMap(
-      migrations.upFiles,
-      (filePattern) => glob.promise(filePattern),
-    ))
-    .sort((a, b) => a.localeCompare(b, undefined, { numeric: true }))
-
-  const downMigrationFiles = (
-    await asyncFlatMap(
-      migrations.downFiles,
-      (filePattern) => glob.promise(filePattern),
-    ))
-    .sort((a, b) => b.localeCompare(a, undefined, { numeric: true }))
-
-  const postMigrationFiles = (
-    await asyncFlatMap(
-      migrations.postMigrationFiles ?? [],
-      (filePattern) => glob.promise(filePattern),
-    ))
-    .sort((a, b) => a.localeCompare(b, undefined, { numeric: true }))
-
-  const migrationTarget = migrations?.target ?? "up/down"
-
-  if (migrationTarget === "up/down"
-    && (
-      // Check if it is not a migration file.
-      upMigrationFiles.filter(file => document.uri.endsWith(file)).length
-      + downMigrationFiles.filter(file => document.uri.endsWith(file)).length === 0
-    )
-  ) {
-    return
-  }
-
-  let shouldContinue = true
-
-  if (shouldContinue) {
-    shouldContinue = await queryMigrations(
-      pgClient, document, downMigrationFiles, logger,
-    )
-  }
-
-  if (shouldContinue) {
-    shouldContinue = await queryMigrations(
-      pgClient, document, upMigrationFiles, logger,
-    )
-  }
-
-  if (shouldContinue) {
-    shouldContinue = await queryMigrations(
-      pgClient, document, postMigrationFiles, logger,
-    )
-  }
-}
-
-async function queryMigrations(
-  pgClient: PostgresClient,
-  document: TextDocument,
-  files: string[],
-  logger: Logger,
-): Promise<boolean> {
-  for await (const file of files) {
-    try {
-      if (document.uri.endsWith(file)) {
-        // allow us to revisit and work on any migration/post-migration file
-        logger.info("Stopping migration execution at the current file")
-
-        return false
-      }
-
-      logger.info(`Migration ${file}`)
-
-      const migration = (await fs.readFile(file, { encoding: "utf8" }))
-        .replace(BEGIN_RE, (m) => "-".repeat(m.length))
-        .replace(COMMIT_RE, (m) => "-".repeat(m.length))
-        .replace(ROLLBACK_RE, (m) => "-".repeat(m.length))
-
-      await pgClient.query(migration)
-    } catch (error: unknown) {
-      const errorMessage = (error as DatabaseError).message
-
-      logger.error(
-        `Stopping migration execution at ${path.basename(file)}: ${errorMessage}`,
-      )
-
-      throw new MigrationError(document, errorMessage)
-    }
-  }
-
-  return true
 }
 
 function queryStatement(
